# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.ResponseHTTP:
    properties:
      data: {}
      details:
        type: string
      success:
        type: boolean
    type: object
  common.ResponsePagination:
    properties:
      data: {}
      details:
        type: string
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
...
  controllers.FeatureDropDown:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
...
  models.Endpoint:
    description: App type information
    properties:
      description:
        type: string
      feature_id:
        type: number
      id:
        type: integer
      method:
        type: string
      name:
        type: string
      route_path:
        type: string
    type: object
...
  models.FeatureGet:
    description: FeatureGet type information
    properties:
      active:
        type: boolean
      description:
        type: string
      endpoints:
        items:
          $ref: '#/definitions/models.Endpoint'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  models.FeaturePost:
    description: FeaturePost type information
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: This is blue-admin API OPENAPI Documentation.
  termsOfService: http://swagger.io/terms/
  title: Swagger blue-admin API
  version: "0.1"
paths:
...
  /endpointfeature/{endpoint_id}:
    delete:
      consumes:
      - application/json
      description: Delete Feature Endpoint
      parameters:
      - description: Feature ID
        in: path
        name: endpoint_id
        required: true
        type: integer
      - description: Endpoint ID
        in: query
        name: feature_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Delete Feature Endpoint
      tags:
      - Features
    patch:
      consumes:
      - application/json
      description: Add Feature to Endpoint
      parameters:
      - description: Endpoint ID
        in: path
        name: endpoint_id
        required: true
        type: integer
      - description: ' Feature ID'
        in: query
        name: feature_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add Feature to Endpoint
      tags:
      - Features
  /feature:
    get:
      consumes:
      - application/json
      description: Get Features
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponsePagination'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.FeatureGet'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      - Refresh: []
      summary: Get Features
      tags:
      - Features
    post:
      consumes:
      - application/json
      description: Add Feature
      parameters:
      - description: Add Feature
        in: body
        name: feature
        required: true
        schema:
          $ref: '#/definitions/models.FeaturePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.FeaturePost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add a new Feature
      tags:
      - Features
  /feature/{feature_id}:
    delete:
      consumes:
      - application/json
      description: Remove feature by ID
      parameters:
      - description: Feature ID
        in: path
        name: feature_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Remove Feature by ID
      tags:
      - Features
    get:
      consumes:
      - application/json
      description: Get feature by ID
      parameters:
      - description: Feature ID
        in: path
        name: feature_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.FeatureGet'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get Feature by ID
      tags:
      - Features
    patch:
      consumes:
      - application/json
      description: Patch Feature
      parameters:
      - description: Patch Feature
        in: body
        name: feature
        required: true
        schema:
          $ref: '#/definitions/models.FeaturePost'
      - description: Feature ID
        in: path
        name: feature_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.FeaturePost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Patch Feature
      tags:
      - Features
  /featuredrop:
    get:
      consumes:
      - application/json
      description: Get FeatureDropDown
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controllers.FeatureDropDown'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get FeatureDropDown
      tags:
      - Feature
...
  /features/{feature_id}:
    put:
      consumes:
      - application/json
      description: Activate/Deactivate Feature
      parameters:
      - description: Feature ID
        in: path
        name: feature_id
        required: true
        type: integer
      - description: Active
        in: query
        name: active
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.FeaturePost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Activate/Deactivate Feature
      tags:
      - Feature
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: X-APP-TOKEN
    type: apiKey
  Refresh:
    description: Description for what is this security definition being used
    in: header
    name: X-REFRESH-TOKEN
    type: apiKey
swagger: "2.0"